fix: fixed /src/components path
