+++ src/app/(user)/page.tsx
@@ -3,7 +3,7 @@ export default function Home() {
   return (
     <main className="flex h-screen flex-col items-start justify-center">
       <div className='text-6xl font-sans text-gray-400 align-top' >Welcome</div>
-      <div className="text-6xl font-bold font-sans">LET'S HUNT EMISSIONS DATA</div>
+      <div className="text-6xl font-bold font-sans">LET&apos;S HUNT EMISSIONS DATA</div>
     </main>
   );
 }
