+++ src/components/hero-section.tsx
@@ -8,7 +8,7 @@ const HeroSection: React.FC = () => {
       <div className="h-[600px]">
         <WelcomeHeader
           text1={"Welcome"}
-          text2="let's hunt emissions data"
+          text2="let&apos;s hunt emissions data"
           className="lg-and-up:pl-20"
         />
         <BodyText className="uppercase text-center text-[12px] leading-[12px] tracking-[0.2em] font-normal">
