+++ src/components/partners-section.tsx
