+++ src/components/styles/welcome-header.tsx
